agaaaaaaaain

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -8,7 +8,7 @@
  */
 hash_table_t *hash_table_create(unsigned long int size)
 {
-	unsigned int i = 0;
+	unsigned  long int i = 0;
 	hash_table_t *ht = malloc(sizeof(hash_table_t));
 
 	if (ht == NULL)
